ViewDistance is now solely controlled by the options menu view distance slider
Former-commit-id: 2c06106723525f9e6718254f63b7ad1ac9058e88

+++ Sample/source/Options.h
@@ -11,7 +11,7 @@ public:
 	float HUDTransparency;
 
 	//View Distance
-	int ViewDistance;
+	float ViewDistance;
 
 	void Print();
 };
